Created CNN Model
Create a Base CNN Model. CNN has needs to be initialized
with a list of inputs, outputs, kernelSizes and
padding(optional and stride(optional). Number of Layers
is based on the number of outChannels in list.Training not
included and CNN not initialized yet.

Also Changed the program run command, plots only show after
-- display is used. Read Updated to show new command

Issue: #37